Update Helper.php and company-form.php

+++ views/api/company-form.php
@@ -23,7 +23,7 @@
 
     <div class="form-group">
         <?php echo $form->labelEx($model, 'quota', ['class' => 'control-label']); ?>
-        <?php echo $form->textField($model, 'quota', ['class' => 'form-control', 'value' => (int)Helper::fileSizeConvert($model['quota']) ? (int)Helper::fileSizeConvert($model['quota']) : 0]); ?>
+        <?php echo $form->textField($model, 'quota', ['class' => 'form-control', 'value' => Helper::fileSizeConvert($model['quota']) ? (int)Helper::fileSizeConvert($model['quota']) : 0]); ?>
         <?php echo $form->error($model, 'quota'); ?>
     </div>
